Support multiple output formats
